ml/requirements.txt
# Core dependencies
flask>=2.0.0
flask-cors>=3.0.0
pandas>=1.3.0
numpy>=1.20.0
yfinance
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.20.0
requests>=2.27.0
beautifulsoup4>=4.10.0

# PDF generation
fpdf>=1.7.2
fpdf2>=2.5.0
reportlab>=3.6.0

# LLM and other dependencies
langchain>=0.0.335
langchain-openai>=0.0.1
langchain-core>=0.1.10
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-google-genai>=0.0.5
langchain-experimental>=0.0.42
openai>=1.0.0
google-generativeai>=0.3.1

# For web searches
duckduckgo-search>=3.0.0